refactor: generate_lance_movesを移動
