+++ packages/rust-core/crates/engine-core/src/movegen/generator/sliding.rs
@@ -170,5 +170,124 @@ fn generate_king_style_moves(
 
 /// Generate lance moves
 pub(super) fn generate_lance_moves(gen: &mut MoveGenImpl, from: Square, promoted: bool) {
-    super::pieces::generate_lance_moves(gen, from, promoted);
+    if promoted {
+        // Promoted lance moves like gold
+        super::pieces::generate_gold_moves(gen, from, promoted);
+        return;
+    }
+
+    let us = gen.pos.side_to_move;
+
+    // Check if lance is at edge and cannot move forward
+    // Black lances move towards rank 0, White lances move towards rank 8
+    // So we need to check if they're already at their destination edge
+    let rank = from.rank();
+    if (us == crate::Color::Black && rank == 0) || (us == crate::Color::White && rank == 8) {
+        return; // Lance at edge cannot move
+    }
+
+    // For lances, we need to manually iterate along the ray in the correct order
+    // because pop_lsb() doesn't guarantee the order we need for blocker detection
+    let mut targets = Bitboard::EMPTY;
+    let file = from.file();
+    let rank = from.rank() as i8;
+
+    // Determine direction and iterate square by square
+    match us {
+        crate::Color::Black => {
+            // Black moves towards rank 0 (up the board)
+            for r in (0..rank).rev() {
+                let sq = Square::new(file, r as u8);
+                if gen.pos.board.all_bb.test(sq) {
+                    // Hit a piece - can capture if enemy, then stop
+                    if !gen.pos.board.occupied_bb[us as usize].test(sq) {
+                        targets.set(sq);
+                    }
+                    break;
+                }
+                targets.set(sq);
+            }
+        }
+        crate::Color::White => {
+            // White moves towards rank 8 (down the board)
+            for r in (rank + 1)..9 {
+                let sq = Square::new(file, r as u8);
+                if gen.pos.board.all_bb.test(sq) {
+                    // Hit a piece - can capture if enemy, then stop
+                    if !gen.pos.board.occupied_bb[us as usize].test(sq) {
+                        targets.set(sq);
+                    }
+                    break;
+                }
+                targets.set(sq);
+            }
+        }
+    }
+
+    // If pinned, can only move along pin ray
+    if gen.pinned.test(from) {
+        let pin_ray = gen.pin_rays[from.index()];
+        let valid_targets = targets & pin_ray;
+        generate_lance_moves_to_targets(gen, from, valid_targets, us);
+        return;
+    }
+
+    // If in check, can only block or capture checker
+    if !gen.checkers.is_empty() {
+        let check_mask = gen.checkers | gen.between_bb(gen.king_sq, gen.checkers.lsb().unwrap());
+        let valid_targets = targets & check_mask;
+        generate_lance_moves_to_targets(gen, from, valid_targets, us);
+        return;
+    }
+
+    // Normal moves
+    generate_lance_moves_to_targets(gen, from, targets, us);
+}
+
+/// Helper function to generate lance moves to specific targets
+fn generate_lance_moves_to_targets(
+    gen: &mut MoveGenImpl,
+    from: Square,
+    targets: crate::Bitboard,
+    us: crate::Color,
+) {
+    use crate::{shogi::Move, PieceType};
+
+    let mut moves = targets;
+    while let Some(to) = moves.pop_lsb() {
+        let captured_type = gen.get_captured_type(to);
+        // Lance must promote if it reaches the last rank
+        let must_promote = match us {
+            crate::Color::Black => to.rank() == 0,
+            crate::Color::White => to.rank() == 8,
+        };
+
+        if must_promote {
+            gen.moves.push(Move::normal_with_piece(
+                from,
+                to,
+                true,
+                PieceType::Lance,
+                captured_type,
+            ));
+        } else {
+            let can_promote = gen.can_promote(from, to, us);
+            gen.moves.push(Move::normal_with_piece(
+                from,
+                to,
+                false,
+                PieceType::Lance,
+                captured_type,
+            ));
+            if can_promote {
+                gen.moves.push(Move::normal_with_piece(
+                    from,
+                    to,
+                    true,
+                    PieceType::Lance,
+                    captured_type,
+                ));
+            }
+        }
+    }
 }
